Model pdf of distance of points in a square grid
This is my first commit, in this repository. In this commit, I model the
probability distribution of the distance of points from the centre of a
square grid of dimension 6 x 6, where the points are chosen randomly,
from with the square grid. I also find out the mean of the distance
random variable. It turns out that the mean is actually the universal
parabolic constant given by the closed form expression sqrt(2) +
log(1+sqrt(2)).

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>