Obtain pi from MonteCarlo runs of buffons needle
In this commit, I have tried to model the Buffons needle experiment to
obtain the value of pi using monte carlo runs. However, the output
doesnt seem to be correct. It is not converging to pi byt to some 3.02.
I need to debug this. The error could be in the modeling as well.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>