Implemented first version of EM algorithm
In this commit, I have added a script which implements the Expectation
Maximization (EM) algorithm. This is a first cut version and is working
under specific cases when the initial estimates are close the ground
truth. However, if the initial estimates are off from the ground truth
values, the algorithm is not converging especially for the variances. I
implemented this algo by following some youtube videos and was still not
very comfortable with the derivation. Now, I have watched the EM
algorithm from Arun Rajkumar from IITM and now have a much clear
understanding of the algorithm. I will attempt to derive it myself and
then implement the code once again.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>