Modified the code with better understanding of the EM algorithm
In this commit, I have made minor modifications to the EM algorithm code
with a better understanding of the algorithm based on the video lectures
of Arun Rajkumar. No mah=jor changes. Now, Im computing the priors also
at each iteration step. But somehow, the priors estimated are still not
converging. Also, if the variance of the GT is changed, then also the
algo is not converging. I need to debug this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>