Generate data points from cluster based on corresponding priors
In this commit, I am generating data points from each cluster based on
the priors for that cluster. Right now, I have coded for a 2 gaussian
mixture model. So for this, if the priors for the 2 gaussian mixtures
are 0.8, 0.2, then, we generate 80 percent of the data points from
cluster 1 and 20 percent of the points from cluster 2. Also, while
plotting the posteriors of the latent variables, Im plotting on the log
scale. But one thing remains to be debugged and that is the convergence
of the prior probabilities. Somehow, the estimated priors are not
matching the programmed priors.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>