Fixed bug in estimating P(x), Priors convering
In this commit, I have fixed a bug becuase of which the priors were not
converging. Previously, mean, variance, etc were all converging to the
ground truth but the priors were not convergin. Now, I have fixed a bug
which was causing this issue. When computing P(x), I should do a weighted
sum of P(x/z) with P(z) (which are the priors). But instead, I was just
doing a sum of P(x/z) across the clsuter dimension. Hence, the P(x) was
not being computed correctly causing the priors to not converge. Now I
have fixed this issue. Now the priors are also converging! This was a
good debug.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>