Moved ML algorithms to a function file
In this commit, I have moved the training, testing functions for each of
the ML algorithms codes so far namely the perceptron, logistic
regression algorithms to a separate file named ml_functions_lib.py.
Gradually, I will remove the separate scripts for perceptron, Logistic
regression and move them to a common script so I can compare the
performace of the algos with the same dataset.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>