Implement batch/online SGD
In this commit, I have implemented the batch and online/stochastic versions of the gradient descent algorithm to solve
the linear least squares problem and compared against the pseudo inverse solution.
I have given a flag to choose between the batch/online version of the gradient descent algorithm to solve
the linear least squares problem.
In the subsequent commits, I will
solve other variants of the least squares and also the LMS algorithm

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>