Verified forward,backward pass,weight updates step
In this commit, I have verified the forward pass, backward pass and the weights updates step for the online/stochastic mode of gradient descent method. I have verified the logic using a reference code available here:
https://towardsdatascience.com/implementing-the-xor-gate-using-backpropagation-in-neural-networks-c1f255b4f20d

My implementation was absolutely correct. The reason I was not getting the correct outputs was because the number of training epochs was insufficient. I had previously dialled in 10k epochs. Now, when I made it to 100k epochs, the output is correct and also the cost function is dying down to 0! So, for stochastic gradient descent, the number of training epochs must be very large since we are using only one training example at a time. But for batch mode/ mini-batch mode, we might not require as many epochs since we are using more than one training sample at a time to update the weights.
Also, the weights initialization with all 0s doesnt seem to be coverging. However, weights initializations from a uniform distribution from [0,1] seems to be converging.
I have tested the NN to predict the output of an XOR gate. Now it is matching the expected ouput. All I had to change was the number of epochs from 10k to 100k

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>