Enable batch mode of gradient descent
In this commit, I have enabled the batch mode of gradient descent(in
addition tot he online/stochastic mode of gradient descent). In the
batch mode of gradient descent, we compute the forward pass and backward
passes parallely for all the test data all at one and the final gradient of the cost function wrt weights is
obtained as the sum/mean of the gradients of each individual test data
point. I have implemented this as a matrix operations.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>