Implement NN on iris data
In this commit, I have successfully tested the NN architecture on the
iris dataset. I was able to achive an accuracy of 96% which is even
beter than what was obtained in the blog. Initally, I was getting poor
perforamance due to the data being biased against the 3rd class in the
dataset while training. i.e the training dataset had very few samples
from class 3 as compred to class 1 and class 2. This was causing the NN
to learn only patterns of class 1 and 2 and not class 3. Now I have
introduced a random shuffling of the data before feeding it to the
training. This helped solve the issue. Also, I plotting both the
training loss and the validation loss/cost functions after each epoch by
 evaluating the cost function on the training and validation data after
obtaining new weights at the end of each epoch. If the model has been
learnt well, we should see both the training loss and validation loss
dropping with epochs. Also, printing the accuracy of the NN on a test
data.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>