Plot the validation loss across epochs
In this commit, I am computing and plotting the validation loss in addition to the traning loss (vs epochs). Validation data is the data which the netrok has not seen before. A network has truly learnt if both the training loss and the validation loss keep decreasing with epochs. If only the training loss decreases but the validation loss increases with epochs, then the model/network has not truly generalized on unseen data and it has probably over fit on the training data.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>