Implement the mini batch version of weight update
In this commit, I have implemented(and tested on the irisi dataset) the mini batch mode of
weights update. With the mini batch and online modes, I was able to
achieve an accuracy of 93.3% accuracy on test data, while with the batch version, I was able to
achieve 96% accuracy on test data. I have also cleaned up the code.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>