ANN for digit classification and achieved accuracy os 90%
In this commit, I have performed digit classification of MNIST dataset and achieved 90% accuracy
on the test data. By trial and erro, I found that an ANN architecture with 1 hidden layer
with 100 neurons and batch mode of GD with stepsize of 1e-6 is required to train the model to achive
90% traning accuracy, validation accuracy and testing accuracy.
Even though, I have given 10000 epochs, the model achieves 90% accuracy after about 1541 epochs.

The input images are 8 bit quantized values and have values ranging from 0 to 255.
One important aspect I learned in this exercise is that, we need to normalize the input image
by 255 to ensure the input data lies between 0 to 1. Normlaizing this way, helps ensure stability
of the weights

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>