Implement Convolutional neural network from scratch
In this commit, I have implemented the Convolutional Neural Network (CNN) from scratch! This includes the backpropagation algorithm for CNNs. I have derived the backpropagation algorithm for the CNNs on papaer all by myself and then implemented the same in the code. The derivation is based on my understanding of the backprop for ANN from Shastry's lectures. I was able to extend the same for CNN backprop. And it worked in the first shot! You do not find the complete derivation of the backprop for CNNs anywhere. Most of the blogs or video tutorials just show for 1 layer with 1 channel. But, I was able to derive it completely! I tested the CNN on the MNIST digit database and was able to achieve a training accuracy of 93% ! I have implemented all the flavours of Gradient descent like, stochastic gradient descent, batch, mini-batch.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>