Test the CNN on MNIST dataset
In this commit, I have tested my implementation of the CNN(including
learning through backprop) on the MNIST dataset. My implementation
worked perfectly! I was able to achieve a training accuracy of 93%! I
used the CNN architecture as given by far1din in his CNN video lecture.
Link:
https://www.youtube.com/watch?v=JboZfxUjLSk&list=PL1sQgSTcAaT7MbcLWacjsqoOQvqzMdUWg&index=1

Need to go from ANN to CNN?
https://www.quora.com/Why-do-we-use-CNN-when-we-already-have-ANN-with-a-fully-connected-structure
1. Exploding number of parameters/weights especially when the size of input image is very large
2. Larger the input size, more the number of weights/parameters --> more number of examples
required to train the network.
3. For an ANN, an image with cat at top left corner of image is different from an image with cat at
bottom right of the image, so it treats it as two different outputs. Whereas, a CNN does a
local weighting/activation and hence for a CNN, both the images are treated the same.

Hence we will move to CNNS for image datasets.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>