GPU kernel for pooling operation
In this commit, I have made a GPU kernel for the maxpool operation. I
have created separate kernels for parallelization across data points as
well as parallelization across kernels. I
bitmatched the former to the CPU implementation and also tested it with
the mnist dataset. The CPU time for each batch (of 256) is 1-2 seconds.
But with the parallelization across datapoints, the time per batch is
0.5 second! Also, the time for each epoch on the CPU was about 6.5
minutes but with the GPU kernel, it is about 2 mins and the total
training and testing on the mnist data was about 1.5 hours. On the CPU it
takes close to 4 hours! This is more than 50% reduction!

I have not fully testd and bitmatched the parallelization across
kernels. I will check that as well. In the next commit, I will
parallelize/kernelize the back prop pooling as well.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>