Kernelized backprop pooling
In this commit, I have put the backprop pooling on the cuda GPU. On the fashion MNIST dataset, with the backprop for the pool layer on GPU, the time per batch has reduced from 3.74s to 2.44 sec i.e about 35% time reduction! On the GPU, before the kernelization of backprop pooling, the time per epoch was projected at about 1.5 hrs. Now, with the kernelization of backprop pooling, the projected time per epoch comes down to 60 minutes or 1 hour! I have also further optimized the backprop pooling logic which has reduced the memory footprint as well as compute time!

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>