Parallelize backprop of polling across kernels
In this commit, I have enabled the backpropagation the pooling
layer on the GPU by parallelizing across the kernels. Previously, I had
done the parallelization across data for the pooling layer back prop.
Now, I have the backprop of the pooling layer parallelized both on
datapoints and kernels. I have also tested and bitmatched the backporp
of pooling layer against the CPU version.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>