Compute training loss and accuracy in batches
Previously, when computing the training loss and accuracy on the entire training data post each epoch, I was giving the entire training data as a big block for a one shot computation of the forward pass on the training data. When the training data was huge like fashion MNIST data where number of training data is about 48000, the size of the input to the convolutional layer goes beyond 4GB which is the size of the RAM available on the GPU. Due to this, the laptop was freezing and also sometimes it would autoshut.
Now, I'm computing the training data output in chunks and not as a bulk, thus reducing the memory being used on the GPU. Currently this feature of data break up has been done for the training data. Next, I will perform the same chunking on the validation data as well.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>