Achieved 85% training and validation accuracy for the Fashion MNIST dataset!
In this commit, finally after nearly 2 months of struggle, I have got a decent accuracy for the fashion MNIST dataset with the CNN! The issue was with the weight initialization! My back propagation is absolutely correct! (previously, I was doubting my backporpagation implementation!) Now, I'm able to achieve about 85% train and validation accuracy with a few epochs! The following the changes which helped in achieveing this.

He initialization instead of normal distribution initialization
He initialization is basically scaling the normal distribution by the sqrt of the nunber of inputs to that layer and sometimes also a multificative gain factor. This gain factor is a function of the activation function used. For example, ReLU has sqrt(2). tanh has 5/3 and so on. Currently, I have scaled based on ReLU only! Need to extend it to other activations as well.

Changed the loss and gradient computation from sum of all datapoints to mean of all datapoints
The loss now is the mean loss of all data points and the gradient is the mean of the gradients of all data points. Now we see reasonable loss values and also the netwrok starts respoding due to more reasonable gradeint values (with sum, gradients blow up)

Implemented the prediction/test also as a batch mode. Now the testing time also is reduced! Previously only the training was parallelized and testing was given as a bulk. Due to this the system was getting stuck or crashing. Now I have parallelized the testing part as well.

Next I will implement the batch normalization to further improve the accuracy

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>