Implemented BN for CNN
In this commit, I have implemented Batch Normalization (BN) for CNN as well! Previously, BN was implemented only for DNN, now I have implemented it for CNN as well. Also, it seems to be working fine and I'm able to obtain about 92%, 90.5 % train and validation accuracy respectvely on the fashion MNIST dataset but with a smaller architecture. This is with BN enabled for CNN and disabled for DNN. This shows the BN for CNN has been implemented correctly!

Also, in this commit, I have fixed a major bug in the CNN update weights function! Previously, the weight update for the kernels was not happening for the last CNN layer. If there were N CNN layers, the weight update loop was running only till N-1 and hence the kernel weights were getting updated for N-1 layers but the last CNN layer kernel weights were not getting updated! Now I have fixed this bug. Somehow, this didn't show up on the fashion MNIST dataset. Only when I was implementing the Batch Normalization and was going through the entire code base for DNN and CNN did I realize this bug. Now I have fixed this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>