Add third argument for BN
In all the dataset loading files, I have added a third argument in the architecture definition which decides whether BN is enabled or not. If 1, BN is enabled and if 0, BN is disabled and it defaults to He weight initialization. BN can be enabled/disabled for only hidden layers. BN is always disabled for input and output layers.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>