Generalized RNN to different length hidden state vector for each layer
In this commit, I have generalized the RNN architecture to cater to different sizes of the hidden state vector
for each RNN layer. Previously, due to matrix only operations, the size of the hidden state vector for each RNN layer
was forced to be same as the input size(also vocab size). But this reduces the flexibility on the hidden states.
By allowing for the hidden state vector to take on any size for each RNN layer, irrespective of the vocab size,
the RNN network can learn different kinds of long term dependencies like " " (quote unquote), etc. Since the hidden state
vector for each layer is different size, we can no longer work with matrices across layers and hence I have
introduced lists across layers. But within a layer, across time steps, we still deal with matrices. However,
the execution time now increases from 1.2 seconds per epoch to 12 seconds per epoch since there are list operations.
The code is running fine. I still need to incorporate the checks for gradient exploding. Moreover, I want to spend some time
understanding and appreciating the problem of exploding gradients in RNNs and not necessarily only from a code crashing stand point.
There's another issue. When I increase the batch size from 1 to say 32, the code still runs fine, but the RNN
doesnt seem to generate meaningful words. This could possibly be because of the following reason.
Since the training data is divided into 32 batches (if batch size is selected to 32), each batch might be starting randomly
and not necessarily from the beginning of a sentence. Hence, the network might not be learning meaningful
long term dependencies. This is my hunch. I need to verify this though! I also need to understand the validation loss and accuracy
for text data.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>