Implement LSTM and back prop
In this commit, I have successfully implemented the Long Short Term Memory (LSTM) network.

I have thoroughly understood and implemented the backpropagation through time for LSTMs.
I will also document my derivation of the backprop for LSTMs in OneNote.
Both RNNs and LSTMs are used to cater to classification and prediction tasks of sequential data.
In RNNs, the gradients are backpropagated through each time step via successive multiplications with
the same Whh matrix and changing diagonal matrix of activation function derivatives.
The major problem of RNNs is the vanishing gradients due to successive multiplications
of the same Whh matrix(with a diagonal matrix of activation function derivatives)
over several time steps (for a given RNN layer). The activation function in RNN is tanh and the derivative of
tanh <= 1. Now, If the eigen values of the Whh matrix are all less than one,
then after several multiplications, the output matrix contains very small values and this leads to
vanishing gradients. I will also attach a small script to illustrate this behaviour.
Now when gradients vanish, the parameter/weights do not get an update and hence the contribution from a
longer context character/word (in the context of language modelling) gets lost.
Hence, an RNN in general, cannot handle longer context windows because
the gradients diminish by the time it propagates to a far off context word or character.
This problem of vanishing gradients in RNNs is handled by making some modifications to the RNN
architecture and this leads to a new architecture called LSTMs or Long Short Term Memory Networks.
LSTMs solve the following problems of RNNs:
    1. Vanishing gradients
    2. Smaller context window for predictions
Note: LSTMs do not solve the exploding gradeints problem! So this issue still needs to be handled with some clipping or normalization!

LSTMs solve the vanishing gradeints problem by getting rid of the successive Whh matrix multiplications.
Instead it provides highway paths for the cell state (which only includes point wise multipliation and additon in its path)
through which the gradients can flow backwards without any diminishing effect.

I have implemented the backprop through time for LSTM. But looks like the code is crashing sometimes due to exploding gradients. I need to clip the gradients. I also need to check for the correctness of the implementation

Following are the changes made in the DNN clss to support LSTMs:
1. Changed name of the update_weights function to compute_gradients_and_update_weights because this function computes both the gradients and also updates the weights using gradeint descent.
2. Storing the gradients of the DNN parameters like weights, baises and batch normalization parameters gamma scaling, beta shift. These gradients are stored for each time step, so that they can be summed up across all time steps to obtain the overall gradient contribution across all time steps.
3. Added new functions compute_gradients and update_weights which are exclusively used for LSTM functions. This just splits the compute_gradients_and_update_weights function into two separate functions.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>